adjust Blurred
