+++ src/components/appNavigation/stackNavigation.js
@@ -12,7 +12,7 @@ import normalize from '../../lib/utils/normalizeText'
 import SideMenuPanel from '../sidemenu/SideMenuPanel'
 import logger from '../../lib/logger/pino-logger'
 import CustomButton, { type ButtonProps } from '../common/buttons/CustomButton'
-import Blurred from '../common/view/Blur/Blurred'
+import Blurred from '../common/view/Blurred'
 import BackButtonHandler from '../../lib/utils/handleBackButton'
 import NavBar from './NavBar'
 import { navigationOptions } from './navigationConfig'
