+++ src/components/common/view/Blurred/index.web.js
